as-cross: Buf関係を修正: txtp, relp

+++ as/cross/as4.c
@@ -4,7 +4,7 @@
 struct Buf {
     short *next; /* next slot */
     short *max;  /* buf max */
-    short addr;  /* seek addr */
+    int   addr;  /* seek addr */
     short data[256];
 };
